refactor: extract useMagCharacterization composable, add replay validation test
